refactor: merging the balances into a singluar util file
